Coffee Bot Simulator
Instructions:
To start playing you must copy and paste the file into a text editor and run the program in your terminal.

Simple yet precise. This coffee bot is a true pioneer. It is still under construction and is far from finished.
In the process of making an App that will be the peak of not only one but many chat bots simulator.